feat(chat): open chat in right drawer from TopNav CTA
