+++ app/src/components/TopNav.tsx
@@ -5,8 +5,8 @@
  *
  * Layout (Obsidian Kinetic, matches Stitch spec):
  *   Left  — KOMPL logo (links to /)
- *   Center — primary nav: Home · Wiki · Chat · Sources
- *   Right  — search archive input (with live dropdown) · settings icon
+ *   Center — primary nav: Home · Wiki · Sources
+ *   Right  — search archive input · Chat (accent CTA) · settings icon
  *
  * Search behaviour:
  *   - Typing ≥ 2 chars fetches /api/pages/search?q=...&limit=3 (250ms debounce)
@@ -22,7 +22,7 @@
 import Link from 'next/link';
 import Image from 'next/image';
 import { usePathname, useRouter, useSearchParams } from 'next/navigation';
-import { Search, Settings } from 'lucide-react';
+import { Search, Settings, MessageSquare } from 'lucide-react';
 import {
   useCallback,
   useEffect,
@@ -31,11 +31,11 @@ import {
   type KeyboardEvent,
 } from 'react';
 import { PAGE_TYPE_VAR } from '../lib/page-type-palette';
+import { useChatDrawer } from './ChatDrawer';
 
 const NAV_ITEMS = [
   { href: '/',     label: 'Home', isActive: (p: string) => p === '/' },
   { href: '/wiki', label: 'Wiki', isActive: (p: string) => p.startsWith('/wiki') },
-  { href: '/chat', label: 'Chat', isActive: (p: string) => p.startsWith('/chat') },
 ] as const;
 
 interface SearchResult {
@@ -59,6 +59,7 @@ export default function TopNav() {
   const router       = useRouter();
   const searchParams = useSearchParams();
   const isAddMode    = searchParams.get('mode') === 'add';
+  const { open: chatOpen, toggle: toggleChat } = useChatDrawer();
 
   const [query,     setQuery]     = useState('');
   const [results,   setResults]   = useState<SearchResult[]>([]);
@@ -365,6 +366,38 @@ export default function TopNav() {
             </div>
           )}
         </div>
+
+        <button
+          type="button"
+          onClick={toggleChat}
+          aria-label="Open chat"
+          aria-pressed={chatOpen}
+          title="Chat"
+          style={{
+            display: 'inline-flex',
+            alignItems: 'center',
+            gap: 8,
+            height: 36,
+            padding: '0 14px',
+            background: chatOpen ? 'var(--accent)' : 'rgba(var(--accent-rgb),0.18)',
+            border: '1px solid var(--accent)',
+            borderRadius: 6,
+            cursor: 'pointer',
+            color: chatOpen ? 'var(--accent-text)' : 'var(--accent)',
+            fontFamily: 'var(--font-heading)',
+            fontSize: 13,
+            fontWeight: 700,
+            letterSpacing: '0.06em',
+            textTransform: 'uppercase',
+            boxShadow: chatOpen
+              ? 'none'
+              : '0 0 0 1px rgba(var(--accent-rgb),0.25), 0 0 18px rgba(var(--accent-rgb),0.2)',
+            transition: 'background var(--transition-fast), color var(--transition-fast), box-shadow var(--transition-fast)',
+          }}
+        >
+          <MessageSquare size={15} strokeWidth={2.25} />
+          Chat
+        </button>
 
         {/* Settings */}
         <Link
